% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smargin.R
\name{print.smargins}
\alias{print.smargins}
\title{Print a smargins object.}
\usage{
\method{print}{smargins}(x)
}
\arguments{
\item{x}{An object of class "smargins".}
}
\description{
Print a smargins object.
}
\author{
Ista Zahn
}
