% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smargin.R
\name{scompare}
\alias{scompare}
\title{Compare expected values at combinations of predictor variable values.}
\usage{
scompare(smargins, var)
}
\arguments{
\item{var}{A character string naming the variable to compute comparisons for.}

\item{object}{An object of class "smargins".}
}
\description{
Compare expected values at combinations of predictor variable values.
}
\author{
Ista Zahn
}
