% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmp.R
\name{scompare_v1}
\alias{scompare_v1}
\title{Compare expected values at combinations of predictor variable values. version 1.}
\usage{
scompare_v1(smargins, var, contrast)
}
\arguments{
\item{var}{A character string naming the variable to compute comparisons for.}

\item{contrast}{A matrix of group contrasts}

\item{object}{An object of class "smargins".}
}
\description{
Compare expected values at combinations of predictor variable values. version 1.
}
\author{
Ista Zahn
}
