% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmp.R
\name{scompare_v2}
\alias{scompare_v2}
\title{Compare expected values at combinations of predictor variable values. version 2.}
\usage{
scompare_v2(smargins, var, contrast.fun)
}
\arguments{
\item{var}{A character string naming the variable to compute comparisons for.}

\item{contrast.fun}{A function that takes expected values for each group and returns the desired QI}

\item{object}{An object of class "smargins".}
}
\description{
Compare expected values at combinations of predictor variable values. version 2.
}
\author{
Ista Zahn
}
