% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmp.R
\name{scompare_v2b}
\alias{scompare_v2b}
\title{Compare expected values at combinations of predictor variable values. version2 b}
\usage{
scompare_v2b(smargins, var, contrast.funs)
}
\arguments{
\item{var}{A character string naming the variable to compute comparisons for.}

\item{contrast.funs}{A list of functions that takes expected values for each group and returns the desired QI}

\item{object}{An object of class "smargins".}
}
\description{
Compare expected values at combinations of predictor variable values. version2 b
}
\author{
Ista Zahn
}
