% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smargin.R
\name{smargins.default}
\alias{smargins.default}
\title{Calculate average marginal effects from a fitted model object.}
\usage{
\method{smargins}{default}(model, ..., n = 5000, coef.fun = coef,
  sim.fun = I, linkinv.fun = family(model)$linkinv, vcov.fun = vcov,
  model.frame.fun = model.frame, model.matrix.fun = function(x, data) {    
  model.matrix(formula(x), data = data) })
}
\arguments{
\item{model}{A fitted model object.}

\item{...}{Named values to set predictor variables to.}

\item{n}{Number of simulations to run.}

\item{coef.fun}{Number of simulations to run.}

\item{sim.fun}{Number of simulations to run.}

\item{linkinv.fun}{Number of simulations to run.}

\item{vcov.fun}{Number of simulations to run.}

\item{model.frame.fun}{Number of simulations to run.}

\item{model.matrix.fun}{Number of simulations to run.}
}
\value{
A data.frame containing predictor variables values and
        expected values of the dependant variable.
}
\description{
Calculate average marginal effects from a fitted model object.
}
\author{
Ista Zahn
}
