% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smargin.R
\name{smargins}
\alias{smargins}
\title{Calculate average marginal effects from a fitted model object.}
\usage{
smargins(model, ..., n = 1000)
}
\arguments{
\item{model}{A fitted model object.}

\item{...}{Further arguments passed to or from other methods.}

\item{n}{Number of simulations to run.}

\item{at}{A named list of values to set predictor variables to.}
}
\value{
A data.frame containing predictor variables values and
    expected values of the dependant variable.
}
\description{
Calculate average marginal effects from a fitted model object.
}
\examples{
library(smargins)
lm.out <- lm(Fertility ~ Education, data = swiss)

smargins(lm.out, at = list(Education = quantile(swiss$Education, c(.25, .50, .75))))

}
\author{
Ista Zahn
}
