% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survreg.R
\name{smargins.survreg}
\alias{smargins.survreg}
\title{Calculate average marginal effects from a fitted model object.}
\usage{
\method{smargins}{survreg}(model, ..., n = 1000)
}
\arguments{
\item{model}{A fitted model object.}

\item{...}{Named values to set predictor variables to.}

\item{n}{Number of simulations to run.}
}
\value{
A data.frame containing predictor variables values and
    expected values of the dependant variable.
}
\description{
Calculate average marginal effects from a fitted model object.
}
\author{
Ista Zahn
}
