% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmp.R
\name{smargins_v2}
\alias{smargins_v2}
\title{Temporary smargins variant}
\usage{
smargins_v2(model, vars, xvalues = NULL, n = 5000, coef.fun = coef,
  sim.fun = I, linkinv.fun = family(model)$linkinv, vcov.fun = vcov,
  model.frame.fun = model.frame, model.matrix.fun = function(x, data) {    
  model.matrix(formula(x), data = data) })
}
\arguments{
\item{model}{A fitted model object.}

\item{vars}{Character vector naming variables}

\item{xvalues}{Named list of values.}

\item{n}{Number of simulations to run.}

\item{coef.fun}{Function to use for computing coefficients.}

\item{sim.fun}{Function to use for adjusting simulations.}

\item{linkinv.fun}{Function to use for cmputing the inverse link.}

\item{vcov.fun}{Function to use for computing the variance-covariance matrix.}

\item{model.frame.fun}{Function to use for extracting the model.frame.}

\item{model.matrix.fun}{Function to use for extracting the model.matrix.}
}
\value{
A data.frame containing predictor variables values and
        expected values of the dependant variable.
}
\description{
Temporary smargins variant
}
\author{
Ista Zahn
}
