% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smargin.R
\name{summary.smargins}
\alias{summary.smargins}
\title{Summarize a smargins object.}
\usage{
\method{summary}{smargins}(object, level = 0.95)
}
\arguments{
\item{object}{An object of class "smargins".}

\item{level}{Level at which confidnece interval should be calculated}
}
\description{
Summarize a smargins object.
}
\author{
Ista Zahn
}
