% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{sumstats}
\alias{sumstats}
\title{Calculate summary statistics}
\usage{
sumstats(x, level = 0.95)
}
\arguments{
\item{x}{A numeric vector, e.g., of expected values}

\item{level}{A fraction between 0 and 1 giving the level at which
to calculate confidence intervals.}
}
\value{
A named vector of summary statistics.
}
\description{
Calculate summary statistics
}
\author{
Ista Zahn
}
